Fix Agena SPS thrust
Using both an engine config and generic RCS on the same part seems to cause some unexpected issues, as the modifiers for the RCS also applies to the engine config. Since I don't wanna mess with the RCS patches, just manually set the engine config back to the correct value.